pyproject.toml
[project]
name = "time-tracking"
version = "0.1.0"
description = "Google Calendar → Rocketlane time tracking automation"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "slack-sdk>=3.0.0",
    "anthropic>=0.40.0",
]

[project.scripts]
pull-my-time-for = "time_tracking.classify_events:main"
post-my-time-for = "time_tracking.post_time_entries:main"
time-tracking-listener = "time_tracking.slack_listener:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/time_tracking"]
